@@ -21,17 +21,17 @@
   "classmap-authoritative": true,
   "prefer-stable": true,
   "require": {
-    "php": ">=7.0",
-    "laminas/laminas-session": "2.9.*",
-    "laminas/laminas-cache": "2.9.*",
-    "laminas/laminas-serializer": "2.9.*",
-    "laminas/laminas-validator": "2.12.*",
-    "laminas/laminas-mail": "2.10.*",
-    "laminas/laminas-i18n": "2.10.*",
+    "php": ">=7.3",
+    "laminas/laminas-session": "2.10.*",
+    "laminas/laminas-cache": "2.11.*",
+    "laminas/laminas-serializer": "2.10.*",
+    "laminas/laminas-validator": "2.14.*",
+    "laminas/laminas-mail": "2.14.*",
+    "laminas/laminas-i18n": "2.11.*",
     "zf1/zend-config": "~1.12",
     "zf1/zend-acl": "~1.12",
-    "zf1/zend-ldap": "~1.12",
-    "zf1/zend-auth": "~1.12",
+    "laminas/laminas-authentication": "2.8.*",
+    "laminas/laminas-ldap": "2.11.0",
     "belhard/zend-db": "dev-master",
     "belhard/xajax": "dev-master",
     "belhard/floatthead": "dev-master",
@@ -49,6 +49,22 @@
     "ext-intl": "*",
     "ext-pdo": "*",
     "ext-xml": "*"
+  },
+  "replace": {
+    "laminas/laminas-cache-storage-adapter-apc": "*",
+    "laminas/laminas-cache-storage-adapter-apcu": "*",
+    "laminas/laminas-cache-storage-adapter-blackhole": "*",
+    "laminas/laminas-cache-storage-adapter-dba": "*",
+    "laminas/laminas-cache-storage-adapter-ext-mongodb": "*",
+    "laminas/laminas-cache-storage-adapter-filesystem": "*",
+    "laminas/laminas-cache-storage-adapter-memcache": "*",
+    "laminas/laminas-cache-storage-adapter-memcached": "*",
+    "laminas/laminas-cache-storage-adapter-mongodb": "*",
+    "laminas/laminas-cache-storage-adapter-redis": "*",
+    "laminas/laminas-cache-storage-adapter-session": "*",
+    "laminas/laminas-cache-storage-adapter-wincache": "*",
+    "laminas/laminas-cache-storage-adapter-xcache": "*",
+    "laminas/laminas-cache-storage-adapter-zend-server": "*"
   },
   "scripts": {
     "post-update-cmd": [
