@@ -728,7 +728,7 @@ class Db {
 	/**
 	 * Получение всех настроек системы
 	 */
-	final private function getAllSettings() {
+	private function getAllSettings() {
 		$key = "all_settings_" . $this->config->database->params->dbname;
 		if (!($this->cache->hasItem($key))) {
 			$res = $this->db->fetchAll("SELECT code, value, is_custom_sw, is_personal_sw FROM core_settings WHERE visible='Y'");
