@@ -1,6 +1,13 @@
-<?
+<?php
+namespace Core2;
+require_once 'Db.php';
 
-class LdapAuth extends Common {
+use Laminas\Authentication\AuthenticationService;
+use Laminas\Authentication\Result;
+use Laminas\Authentication\Adapter\Ldap as LdapAdapter;
+use Laminas\Ldap\Ldap;
+
+class LdapAuth extends Db {
 	const ST_LDAP_AUTH_SUCCESS 		= 1;
 	const ST_LDAP_USER_NOT_FOUND	= 2;
 	const ST_LDAP_INVALID_PASSWORD	= 3;
@@ -47,44 +54,50 @@ class LdapAuth extends Common {
 	 * LDAP auth only
 	 * @param $login
 	 * @param $password
-	 * @throws Exception
+	 * @throws \Exception
 	 */
 	public function auth($login, $password) {
 		try {
-			$auth = Zend_Auth::getInstance();
-			$config = Zend_Registry::getInstance()->get('config');
+			$auth = new AuthenticationService();
+			$config = \Zend_Registry::getInstance()->get('config');
 			$log_path = $config->ldap->log_path;
 			$root = $config->ldap->root;
 			$admin = $config->ldap->admin;
+			$role_id = $config->ldap->role_id;
 			$options = $config->ldap->toArray();
 			unset($options['active']);
 			unset($options['log_path']);
 			unset($options['root']);
 			unset($options['admin']);
-			$adapter = new Zend_Auth_Adapter_Ldap($options, $login, $password);
+			unset($options['role_id']);
+			$adapter = new LdapAdapter($options, $login, $password);
 			$result = $auth->authenticate($adapter);
 			if (!$result->isValid()) {
 				// Authentication failed; print the reasons why
-				switch ($result->getCode()) {
+                $code = $result->getCode();
+				switch ($code) {
 
-					case Zend_Auth_Result::FAILURE_IDENTITY_NOT_FOUND:
+					case Result::FAILURE_IDENTITY_NOT_FOUND:
 						$this->setStatus(LdapAuth::ST_LDAP_USER_NOT_FOUND);
 						break;
 
-					case Zend_Auth_Result::FAILURE_CREDENTIAL_INVALID:
+					case Result::FAILURE_CREDENTIAL_INVALID:
 						$this->setStatus(LdapAuth::ST_LDAP_INVALID_PASSWORD);
 						break;
 
 				}
 
 				$msg = '';
-				foreach ($result->getMessages() as $message) {
+				foreach ($result->getMessages() as $i => $message) {
+                    if ($i < 2) continue; // Messages from position 2 and up are informational messages from the LDAP
+
 					$msg .= "$message\n";
 				}
-				throw new Exception($msg);
-			} else {
+				throw new \Exception($msg);
+			}
+			else {
 				if ($result->getIdentity() !== $auth->getIdentity()) {
-					throw new Exception('Ошибка аутентификации');
+					throw new \Exception('Ошибка аутентификации');
 				}
 				$this->setStatus(LdapAuth::ST_LDAP_AUTH_SUCCESS);
 				$userData = array('login' => $result->getIdentity());
@@ -96,10 +109,12 @@ class LdapAuth extends Common {
 				if ($admin === $userData['login']) {
 					$userData['admin'] = true;
 				}
+				//$ldapInfo = $this->getLdapInfo($userData['login']);
+                $userData['role_id'] = $role_id;
 				$this->setUserData($userData);
 			}
 
-		} catch (Exception $e) {
+		} catch (\Exception $e) {
 			if (!$this->status) $this->setStatus(LdapAuth::ST_ERROR);
 			$this->setMessage($e->getMessage());
 		}
@@ -110,12 +125,13 @@ class LdapAuth extends Common {
      */
 	public function getLdapInfo($login) {
 
-		$config = Zend_Registry::getInstance()->get('config');
+		$config = \Zend_Registry::getInstance()->get('config');
 		$options = $config->ldap->toArray();
 		unset($options['active']);
 		unset($options['log_path']);
 		unset($options['root']);
 		unset($options['admin']);
+		unset($options['role_id']);
 		$options = current($options);
 		//echo "<PRE>";print_r($options);echo "</PRE>";die;
 		//$options['accountCanonicalForm'] = 2;
@@ -127,7 +143,7 @@ class LdapAuth extends Common {
 
 		$key = 'profile_' . md5($login);
 		if (!($this->cache->test($key))) {
-			$ldap = new Zend_Ldap($options);
+			$ldap = new Ldap($options);
 			$ldap->bind();
 			$temp = explode('\\', $login);
 			if (isset($temp[1])) $login = $temp[1];
@@ -156,8 +172,8 @@ class LdapAuth extends Common {
 					'firstname' => $data['givenname'][0]
 				));
 			}
-			$this->auth->LN = $data['sn'][0];
-			$this->auth->FN = $data['givenname'][0];
+			//$this->auth->LN = $data['sn'][0];
+			//$this->auth->FN = $data['givenname'][0];
 			$this->cache->save($login, $key, array('core_users'));
 		}
 	}
